fix hover behavior
