+++ src/components/NavMenu.tsx
@@ -1,4 +1,4 @@
-import { type FC } from 'react';
+import { useState, useRef, type FC } from 'react';
 import * as NavigationMenu from '@radix-ui/react-navigation-menu';
 import cx from 'classnames';
 
@@ -8,7 +8,7 @@ interface NavMenuOption {
 }
 
 interface Props {
-	isOpen: boolean;
+	id: number;
 	currentPage: string;
 	options: NavMenuOption[];
 	selectedPage: string;
@@ -16,19 +16,39 @@ interface Props {
 }
 
 const NavMenu: FC<Props> = ({
-	isOpen,
+	id,
 	currentPage,
 	options,
 	selectedPage,
 	selectedPageDisplayName,
 }) => {
+	const [openMenu, setOpenMenu] = useState<number | null>(null);
+	const closeTimeoutRef = useRef<NodeJS.Timeout | null>(null);
+	const handleMouseEnter = (id: number) => {
+		if (closeTimeoutRef.current) {
+			clearTimeout(closeTimeoutRef.current);
+			closeTimeoutRef.current = null;
+		}
+		setOpenMenu(id);
+	};
+
+	const handleMouseLeave = () => {
+		// Set a delay before closing the menu
+		closeTimeoutRef.current = setTimeout(() => {
+			setOpenMenu(null);
+		}, 300); // 300ms delay before closing
+	};
 	return (
 		<NavigationMenu.Root
 			orientation='vertical'
 			className='flex relative justify-start'
 		>
 			<NavigationMenu.List className='flex flex-column'>
-				<NavigationMenu.Item>
+				<NavigationMenu.Item
+					key={id}
+					onMouseEnter={() => handleMouseEnter(id)}
+					onMouseLeave={handleMouseLeave}
+				>
 					<NavigationMenu.Trigger
 						className={cx(
 							'flex flex-row gap-1 justify-between items-center',
@@ -38,7 +58,7 @@ const NavMenu: FC<Props> = ({
 						)}
 					>
 						{selectedPageDisplayName}
-						{isOpen ? (
+						{openMenu === id ? (
 							<img
 								src='/arrow-up.svg'
 								height={16}
@@ -54,7 +74,7 @@ const NavMenu: FC<Props> = ({
 							/>
 						)}
 					</NavigationMenu.Trigger>
-					{isOpen && (
+					{openMenu === id && (
 						<NavigationMenu.Content className='flex flex-column py-4 px-2 rounded-3xl bg-mwasi border border-otherGrey2 absolute top-8 left-0 w-[214px]'>
 							<ul className='w-full'>
 								{options.map((option) => {
